Creando los archivos TecnicoUiState y TicketUiState
